+++ src/runtime/parallel/worker.ts
@@ -3,7 +3,6 @@ import { IdGenerator } from '@cucumber/messages'
 import { duration } from 'durations'
 import { EventEmitter } from 'events'
 import { pathToFileURL } from 'url'
-import StackTraceFilter from '../../stack_trace_filter'
 import supportCodeLibraryBuilder from '../../support_code_library_builder'
 import { ISupportCodeLibrary } from '../../support_code_library_builder/types'
 import { doesHaveValue } from '../../value_checker'
@@ -33,7 +32,6 @@ export default class Worker {
   private filterStacktraces: boolean
   private readonly newId: IdGenerator.NewId
   private readonly sendMessage: IMessageSender
-  private readonly stackTraceFilter: StackTraceFilter
   private supportCodeLibrary: ISupportCodeLibrary
   private worldParameters: any
   private runTestRunHooks: RunsTestRunHooks
@@ -55,7 +53,6 @@ export default class Worker {
     this.exit = exit
     this.sendMessage = sendMessage
     this.eventBroadcaster = new EventEmitter()
-    this.stackTraceFilter = new StackTraceFilter()
     this.eventBroadcaster.on('envelope', (envelope: messages.Envelope) => {
       this.sendMessage({
         jsonEnvelope: JSON.stringify(envelope),
@@ -81,9 +78,6 @@ export default class Worker {
 
     this.worldParameters = options.worldParameters
     this.filterStacktraces = filterStacktraces
-    if (this.filterStacktraces) {
-      this.stackTraceFilter.filter()
-    }
     this.runTestRunHooks = makeRunTestRunHooks(
       options.dryRun,
       this.supportCodeLibrary.defaultTimeout,
@@ -102,9 +96,6 @@ export default class Worker {
       this.supportCodeLibrary.afterTestRunHookDefinitions,
       'an AfterAll'
     )
-    if (this.filterStacktraces) {
-      this.stackTraceFilter.unfilter()
-    }
     this.exit(0)
   }
 
@@ -137,6 +128,7 @@ export default class Worker {
       testCase,
       retries,
       skip,
+      filterStackTraces: this.filterStacktraces,
       supportCodeLibrary: this.supportCodeLibrary,
       worldParameters: this.worldParameters,
     })